@@ -1,11 +1,15 @@
-﻿using SaberFactory.Editor;
+﻿using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+using FlowUi.Helpers;
+using SaberFactory.Editor;
 using SaberFactory.Helpers;
 using SaberFactory.Models;
 using SaberFactory.UI;
-using SaberFactory.UI.CustomSaber;
+using SaberFactory.UI.Flow;
 using SaberFactory.UI.Lib;
-using SaberFactory.UI.Lib.BSML;
 using SiraUtil;
+using UnityEngine;
 using Zenject;
 
 namespace SaberFactory.Installers
@@ -14,23 +18,27 @@ namespace SaberFactory.Installers
     {
         public override void InstallBindings()
         {
-            Container.Bind<EditorInstanceManager>().AsSingle();
+            Container.Bind<SaberInstanceManager>().AsSingle();
 
-            MainUIInstaller.Install(Container);
+            //MainUIInstaller.Install(Container);
 
             BindRemoteSabers();
 
-            Container.Bind<BaseUiComposition>().To<CustomSaberUiComposition>().AsSingle();
+            //Container.Bind<BaseUiComposition>().To<CustomSaberUiComposition>().AsSingle();
             Container.Bind<SaberGrabController>().AsSingle();
-            Container.BindInterfacesAndSelfTo<Editor.Editor>().AsSingle();
+            //Container.BindInterfacesAndSelfTo<Editor.Editor>().AsSingle();
 
-            Container.BindInterfacesAndSelfTo<SaberFactoryMenuButton>().AsSingle();
+            //Container.BindInterfacesAndSelfTo<SaberFactoryMenuButton>().AsSingle();
 
             Container.Bind<TrailPreviewer>().AsSingle();
 
             Container.Bind<MenuSaberProvider>().AsSingle();
 
             Container.BindInterfacesAndSelfTo<GizmoAssets>().AsSingle();
+            
+            Container.BindFromAssetBundle<Prefab.Editor>("Resources.Prefabs.SaberFactoryEditor");
+            Container.BindFlowUi<SaberFactoryMainUi>().CreateMenuButton("Saber Factory", "Sabers!");
+            Container.Bind<SaberListController>().AsTransient();
 
 #if DEBUG
             //Container.Bind<DebugMenu>().FromNewComponentOnNewGameObject().AsSingle().NonLazy();
