@@ -10,28 +10,13 @@ namespace SaberFactory
     internal static class ResharperPostfix
     {
         [SourceTemplate]
-        public static void isnull(this object obj)
-        {
-            if (obj is null)
-            {
-                //$ $END$
-            }
-        }
-
-        [SourceTemplate]
-        public static void isnotnull(this object obj)
-        {
-            if (obj is { })
-            {
-                //$ $END$
-            }
-        }
-
-        [SourceTemplate]
         public static void retnull(this object obj)
         {
-            if (obj is null)
-            { }
+            if (obj == null)
+            {
+                return;
+            }
+            
             //$ $END$
         }
     }
